+++ src/main/java/com/example/smd/services/MaterialService.java
@@ -33,12 +33,19 @@ public class MaterialService {
 
     // 1. Create
     @Transactional
-    public MaterialResponse create(MaterialRequest request) {
+    public MaterialResponse create(MaterialRequest request, String accountId) {
+
+        var account = accountService.getAccountById(accountId);
+        String roleName = account.getRole().getRoleName();
+        if (!(roleName.equals("COLLABORATOR") || roleName.equals("PDCM"))) {
+            throw new AppException(ErrorCode.ACCESS_DENIED_FOR_ROLE);
+        }
+
         Syllabus syllabus = syllabusRepository.findById(request.getSyllabusId())
                 .orElseThrow(() -> new AppException(ErrorCode.SYLLABUS_NOT_FOUND));
 
-        if(!(syllabus.getStatus().equals("DRAFT") || syllabus.getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString()))) {
-            throw new AppException(ErrorCode.SYLLABUS_NOT_EDITABLE);
+        if (!(syllabus.getStatus().equals("DRAFT") || syllabus.getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString()))) {
+            throw new AppException(ErrorCode.MATERIAL_CANNOT_CREATE);
         }
 
         Material material = materialMapper.toEntity(request);
@@ -51,11 +58,17 @@ public class MaterialService {
 
     // 2. Update
     @Transactional
-    public MaterialResponse update(UUID id, MaterialRequest request) {
+    public MaterialResponse update(UUID id, MaterialRequest request, String accountId) {
+        var account = accountService.getAccountById(accountId);
+        String roleName = account.getRole().getRoleName();
+        if (!(roleName.equals("COLLABORATOR") || roleName.equals("PDCM"))) {
+            throw new AppException(ErrorCode.ACCESS_DENIED_FOR_ROLE);
+        }
+
         Material material = materialRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.MATERIAL_NOT_FOUND));
 
-        if(!(material.getStatus().equals("DRAFT") || material.getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString()))) {
+        if (!(material.getStatus().equals("DRAFT") || material.getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString()))) {
             throw new AppException(ErrorCode.MATERIAL_NOT_EDITABLE);
         }
 
@@ -83,24 +96,27 @@ public class MaterialService {
 
     // 4. Delete
     @Transactional
-    public void delete(UUID id) {
-        try {
-            // Kiểm tra xem Material có tồn tại không
-            Material material = materialRepository.findById(id)
-                    .orElseThrow(() -> new AppException(ErrorCode.MATERIAL_NOT_FOUND));
+    public void delete(UUID id, String accountId) {
 
-            if(!(material.getSyllabus().getStatus().equals("DRAFT") || material.getSyllabus().getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString()))) {
-                throw new AppException(ErrorCode.SYLLABUS_NOT_EDITABLE);
-            }
+        var account = accountService.getAccountById(accountId);
+        String roleName = account.getRole().getRoleName();
+        if (!(roleName.equals("COLLABORATOR") || roleName.equals("PDCM"))) {
+            throw new AppException(ErrorCode.ACCESS_DENIED_FOR_ROLE);
+        }
 
-            if(material.getStatus().equals("DRAFT")) {
-                materialRepository.delete(material);
-            } else{
-                material.setStatus("ARCHIVED");
-                materialRepository.save(material);
-            }
-        } catch (IllegalArgumentException e) {
-            throw new AppException(ErrorCode.UNCATEGORIZED_EXCEPTION);
+        // Kiểm tra xem Material có tồn tại không
+        Material material = materialRepository.findById(id)
+                .orElseThrow(() -> new AppException(ErrorCode.MATERIAL_NOT_FOUND));
+
+        if (!(material.getSyllabus().getStatus().equals("DRAFT") || material.getSyllabus().getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString()))) {
+            throw new AppException(ErrorCode.MATERIAL_NOT_EDITABLE);
+        }
+
+        if (material.getStatus().equals("DRAFT")) {
+            materialRepository.delete(material);
+        } else {
+            material.setStatus("ARCHIVED");
+            materialRepository.save(material);
         }
     }
 
@@ -121,7 +137,15 @@ public class MaterialService {
 
         // 3. Ép buộc Role thấp chỉ được xem PUBLISHED
         if (roleName.equals("STUDENT") || roleName.equals("LECTURER")) {
-            finalStatus = "PUBLISHED";
+            if (!finalStatus.equals(SyllabusStatus.PUBLISHED.toString())) {
+                throw new AppException(ErrorCode.ACCESS_DENIED_FOR_ROLE);
+            }
+        }
+
+        if (finalStatus.equals(PloStatus.DRAFT.toString())) {
+            if (!(account.getRole().getRoleName().equals("PDCM") || account.getRole().getRoleName().equals("COLLABORATOR"))) {
+                throw new AppException(ErrorCode.ACCESS_DENIED_FOR_ROLE);
+            }
         }
 
         // 4. Truy vấn dữ liệu
@@ -179,8 +203,8 @@ public class MaterialService {
             }
         }
 
-        if(material.getStatus().equals("DRAFT")) {
-            if(!(account.getRole().getRoleName().equals("PDCM") || account.getRole().getRoleName().equals("COLLABORATOR"))){
+        if (material.getStatus().equals("DRAFT") || material.getStatus().equals(SyllabusStatus.REVISION_REQUESTED.toString())) {
+            if (!(account.getRole().getRoleName().equals("PDCM") || account.getRole().getRoleName().equals("COLLABORATOR"))) {
                 throw new AppException(ErrorCode.ACCESS_DENIED_FOR_ROLE);
             }
         }
